pyproject.toml
[project]
name = "crmarena"
version = "0.1.0"
description = "CRMArena: Understanding the Capacity of LLM Agents to Perform Professional CRM Tasks in Realistic Environments"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "attrs==24.2.0",
    "certifi==2024.7.4",
    "cffi==1.17.0",
    "charset-normalizer==3.3.2",
    "cryptography==43.0.0",
    "distro==1.9.0",
    "datasets",
    "exceptiongroup==1.2.2",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httpx==0.27.0",
    "idna==3.7",
    "isodate==0.6.1",
    "jiter==0.5.0",
    "lxml==5.2.2",
    "more-itertools==10.4.0",
    "numpy==2.0.1",
    "openai>=1.68.2",
    "pandas==2.2.2",
    "platformdirs==4.2.2",
    "pycparser==2.22",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "PyJWT==2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-toolbelt==1.0.0",
    "simple-salesforce==1.12.6",
    "six==1.16.0",
    "sniffio==1.3.1",
    "tenacity==9.0.0",
    "tqdm==4.66.5",
    "typing_extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "zeep==4.2.1",
    "sacrebleu==2.4.3",
    "rouge==1.0.1",
    "litellm==1.67.2",
    "together==1.3.1",
]

[project.urls]
Homepage = "https://github.com/SalesforceAIResearch/CRMArena"
Repository = "https://github.com/SalesforceAIResearch/CRMArena"
"Bug Reports" = "https://github.com/SalesforceAIResearch/CRMArena/issues"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.hatch.build.targets.wheel]
packages = ["crm_sandbox"]
